chore: moved express type to own file in types/express.d.ts
